# api/openapi/v1/keyip.yaml
openapi: "3.0.3"
info:
  title: KeyIP-Intelligence API
  version: "1.0.0"
  description: |
    KeyIP-Intelligence is an AI-driven intellectual property lifecycle management platform
    designed for the OLED organic materials industry. It integrates molecular-level chemical
    understanding with patent law intelligence to support patent mining, infringement
    monitoring, portfolio optimization, and lifecycle management.

    ## Authentication

    All API endpoints require Bearer token authentication unless otherwise specified.
    Tokens are issued via the configured identity provider (Keycloak / OIDC-compatible).

    ## Pagination

    List endpoints support page-based pagination via `page` and `page_size` query parameters.
    Responses include pagination metadata.

    ## Error Handling

    All errors follow a unified structure:
    ```json
    {
      "code": "NOT_FOUND",
      "message": "The specified resource does not exist"
    }
    ```
  contact:
    name: KeyIP-Intelligence Open Source Community
    url: https://github.com/turtacn/KeyIP-Intelligence
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.keyip.example.com/api/v1
    description: Production server
  - url: https://staging-api.keyip.example.com/api/v1
    description: Staging server
  - url: http://localhost:8080
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/portfolios:
    get:
      tags: [Portfolios]
      summary: List portfolios
      operationId: listPortfolios
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Portfolio list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioListResponse"

    post:
      tags: [Portfolios]
      summary: Create portfolio
      operationId: createPortfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioRequest"
      responses:
        "201":
          description: Portfolio created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/portfolios/{id}:
    get:
      tags: [Portfolios]
      summary: Get portfolio
      operationId: getPortfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Portfolio detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Portfolios]
      summary: Update portfolio
      operationId: updatePortfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePortfolioRequest"
      responses:
        "200":
          description: Portfolio updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Portfolios]
      summary: Delete portfolio
      operationId: deletePortfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Portfolio deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/portfolios/{id}/patents:
    post:
      tags: [Portfolios]
      summary: Add patents to portfolio
      operationId: addPortfolioPatents
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioPatentsRequest"
      responses:
        "200":
          description: Patents added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Portfolios]
      summary: Remove patents from portfolio
      operationId: removePortfolioPatents
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioPatentsRequest"
      responses:
        "200":
          description: Patents removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/portfolios/{id}/analysis:
    get:
      tags: [Portfolios]
      summary: Get portfolio analysis
      operationId: getPortfolioAnalysis
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Portfolio analysis
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/portfolios/{id}/valuation:
    get:
      tags: [Portfolios]
      summary: Get portfolio valuation
      operationId: getPortfolioValuation
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Portfolio valuation
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/portfolios/{id}/valuation/run:
    post:
      tags: [Portfolios]
      summary: Run portfolio valuation
      operationId: runPortfolioValuation
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Valuation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  portfolio_id:
                    type: string
                  total_value:
                    type: number
                  by_jurisdiction:
                    type: object
                  by_status:
                    type: object
                  recommendations:
                    type: array
                    items:
                      type: object
                  message:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/portfolios/{id}/gap-analysis:
    get:
      tags: [Portfolios]
      summary: Get portfolio gap analysis
      operationId: getPortfolioGapAnalysis
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Gap analysis
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/portfolios/{id}/gap-analysis/run:
    post:
      tags: [Portfolios]
      summary: Run portfolio gap analysis
      operationId: runPortfolioGapAnalysis
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Gap analysis result
          content:
            application/json:
              schema:
                type: object
                properties:
                  portfolio_id:
                    type: string
                  total_patents:
                    type: integer
                  gap_findings:
                    type: array
                    items:
                      type: object
                  message:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/portfolios/{id}/optimize:
    post:
      tags: [Portfolios]
      summary: Optimize portfolio
      description: Runs portfolio optimization and returns recommendations.
      operationId: optimizePortfolio
      parameters:
        - $ref: "#/components/parameters/PortfolioId"
      responses:
        "200":
          description: Optimization results
          content:
            application/json:
              schema:
                type: object
                properties:
                  portfolio_id:
                    type: string
                  total_patents:
                    type: integer
                  optimization_suggestions:
                    type: array
                    items:
                      type: object
                  estimated_value:
                    type: number
                  message:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Lifecycle
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      description: Page number (starts at 1).
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: page_size
      in: query
      description: Number of items per page (default 20, max 100).
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
    PortfolioId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error description.

    # -------------------------------------------------------------------------
    # Health
    # -------------------------------------------------------------------------
...
    CreatePortfolioRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        patent_ids:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
    UpdatePortfolioRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    PortfolioResponse:
      type: object
      description: Portfolio detail. Shape determined by the service layer.
    PortfolioListResponse:
      type: object
      description: Paginated portfolio list. Shape determined by the service layer.
    PortfolioPatentsRequest:
      type: object
      required: [patent_ids]
      properties:
        patent_ids:
          type: array
          items:
            type: string

    # -------------------------------------------------------------------------
    # Lifecycle
    # -------------------------------------------------------------------------
...